terraform: remove node module file
